requirements_ai.txt
Django==5.2.7
djangorestframework==3.15.2

# Data import and processing
pandas>=2.0.0
xport>=3.3.0  # For NHANES XPT file reading
openpyxl>=3.1.0  # For Excel files (optional)

# ML and scientific computing
numpy>=1.24.0
scikit-learn>=1.3.0

# Data validation
python-dateutil>=2.8.0
